Promote cardano_block_scanner module to a directory
So all types defined in it can be split to several files.

+ rework tests so they are specialized on the part they address: either
  the scanner or the streamer.
